restructure computation to decouple reporting and analysing.
in brief:

a channel parameterised by the analysis type, can be in three states. a
result, complete or reported.

* result
an issue that came back from shellcheck.

* complete
all scripts have been analysed but not necessarily reported.

* reported
all scripts have been analysed and reported.

previously we were accumulating all of the results before iterating and
printing them. this could come become very unsatisfactory when we have
large repos with many hundreds of scripts for analysis.