Run in alpine with glibc/gmp and a proper locale
The hGetContents exception which we tried to fix by setting LANG was not
actually fixed. I believe this is because busybux ships with no locale
support and so setting LANG had no effect.

To fix this, I first moved to alpine which is quickly replacing busybox
as the go-to runtime docker base image. The fix then required
installation of an alternate glibc package to define the locale and
allowed for installation of the gmp package instead of copying the
library manually.

The image size increases from ~26MB to ~56MB with this change.