Emit content-based fingerprints
This engine doesn't currently output fingerprints, it relies on the
default fingerprinting algorithm present in the CLI:

    md5(file-name + check-name)

This algorithm is naive and doesn't work very well; there are plans to
improve it. The problem is that any issues found by the same check in
the same file will receive the same fingerprint.

As an example, pbrisbin/aurget has 64 (conceptually) unique issues but
only 7 unique fingerprints. Ultimately, this results in issues not
appearing as fixed or new when they should (because the list of *unique*
fingerprints may not change)

This engine was chosen as a pilot for a new fingerprinting algorithm. If
it goes well, something like what's implemented here will become the CLI
default when an engine does not emit its own fingerprints.

This engine was chosen for the following reasons:

- Shellcheck is more susceptible to the failings of our current
  algorithm due to the nature of shell scripts (often one big file with
  many issues found by the same check)
- While not "official", it is used often enough that we can see the
  results of our approach more than if we chose a less fequently-used
  engine

The algorithm uses the content of the issue as additional entropy into
the fingerprint:

    md5(file-name + check-name + content)

Where content is the line or source code where the issue appears,
stripped of all whitespace (whitespace changes should not affect
fingerprint).

The entire line is read (ignoring column information) because otherwise
many issues would still receive the same fingerprint. Consider a source
file with the two lines "foo = $*" and "bar = $*". These are distinct
issues but would receive the same fingerprint if derived using column
information (which would get you the same unquoted "$*" part in both).